Revert "perf: update reconcile to only calculate on changes"
This reverts commit f805689248118139717802655271227b6ff5dfb0.
