+++ internal/controller/clusterresourcequota_controller.go
@@ -3,6 +3,7 @@ package controller
 import (
 	"context"
 	"fmt"
+	"slices"
 	"sort"
 	"strings"
 	"time"
@@ -17,18 +18,115 @@ import (
 	"github.com/powerhome/pac-quota-controller/pkg/kubernetes/storage"
 	"github.com/powerhome/pac-quota-controller/pkg/kubernetes/usage"
 	"github.com/powerhome/pac-quota-controller/pkg/metrics"
+	"github.com/prometheus/client_golang/prometheus"
 	"go.uber.org/zap"
+	appsv1 "k8s.io/api/apps/v1"
+	autoscalingv1 "k8s.io/api/autoscaling/v1"
+	batchv1 "k8s.io/api/batch/v1"
 	corev1 "k8s.io/api/core/v1"
+	networkingv1 "k8s.io/api/networking/v1"
 	apiequality "k8s.io/apimachinery/pkg/api/equality"
 	"k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/labels"
 	"k8s.io/apimachinery/pkg/runtime"
+	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/client-go/kubernetes"
 	ctrl "sigs.k8s.io/controller-runtime"
+	"sigs.k8s.io/controller-runtime/pkg/builder"
 	"sigs.k8s.io/controller-runtime/pkg/client"
+	"sigs.k8s.io/controller-runtime/pkg/event"
+	"sigs.k8s.io/controller-runtime/pkg/handler"
+	"sigs.k8s.io/controller-runtime/pkg/predicate"
+	"sigs.k8s.io/controller-runtime/pkg/reconcile"
 )
+
+// resourceUpdatePredicate implements a custom predicate function to filter resource updates.
+// It's designed to trigger reconciliation only on meaningful changes, such as spec updates
+// or pod phase changes to/from terminal states, while ignoring noisy status-only updates.
+// Pods going from pending to running or from running to pending
+// are not considered terminal and do not trigger reconciliation.
+// AKA should be accounted for the resource usage
+type resourceUpdatePredicate struct {
+	predicate.Funcs
+}
+
+// Update implements the update event filter.
+func (resourceUpdatePredicate) Update(e event.UpdateEvent) bool {
+	if e.ObjectOld == nil || e.ObjectNew == nil {
+		// Invalid event, ignore
+		return false
+	}
+
+	// Always reconcile if the object's generation changes (i.e., spec was updated).
+	if e.ObjectNew.GetGeneration() != e.ObjectOld.GetGeneration() {
+		return true
+	}
+
+	// Special handling for Pods: reconcile if the pod transitions to or from a terminal state
+	// or if there's a significant status change (like an init container finishing).
+	if podOld, ok := e.ObjectOld.(*corev1.Pod); ok {
+		if podNew, ok := e.ObjectNew.(*corev1.Pod); ok {
+			// Trigger on terminal state transition
+			if pod.IsPodTerminal(podOld) != pod.IsPodTerminal(podNew) {
+				return true
+			}
+			// Trigger if any container (init or app) has terminated since last update
+			if containerTerminated(podOld.Status.InitContainerStatuses, podNew.Status.InitContainerStatuses) ||
+				containerTerminated(podOld.Status.ContainerStatuses, podNew.Status.ContainerStatuses) {
+				return true
+			}
+		}
+	}
+
+	// For all other cases, if the generation hasn't changed, ignore the update event.
+	// This prevents reconciliation loops caused by the controller's own status updates on the CRQ
+	// or other insignificant status changes on watched resources.
+	return false
+}
+
+// Delete implements the delete event filter.
+func (resourceUpdatePredicate) Delete(e event.DeleteEvent) bool {
+	if e.Object == nil {
+		// Invalid event, ignore
+		return false
+	}
+
+	// Trigger reconciliation on pod deletions
+	if _, ok := e.Object.(*corev1.Pod); ok {
+		return true
+	}
+
+	return false
+}
+
+// containerTerminated returns true if any container in the new statuses has transitioned to Terminated
+// while it was not terminated in the old statuses, or if the set of containers has changed.
+func containerTerminated(oldStatuses, newStatuses []corev1.ContainerStatus) bool {
+	if len(oldStatuses) != len(newStatuses) {
+		return true
+	}
+
+	oldTerminated := make(map[string]bool)
+	oldExists := make(map[string]bool)
+	for _, s := range oldStatuses {
+		oldExists[s.Name] = true
+		if s.State.Terminated != nil {
+			oldTerminated[s.Name] = true
+		}
+	}
+
+	for _, s := range newStatuses {
+		if !oldExists[s.Name] {
+			return true
+		}
+		if s.State.Terminated != nil && !oldTerminated[s.Name] {
+			return true
+		}
+	}
+	return false
+}
 
 // ClusterResourceQuotaReconciler reconciles a ClusterResourceQuota object
 type ClusterResourceQuotaReconciler struct {
@@ -47,7 +145,19 @@ type ClusterResourceQuotaReconciler struct {
 	ExcludedNamespaces       []string
 	// previousNamespacesByQuota tracks namespaces from previous reconciliation for change detection
 	previousNamespacesByQuota map[string][]string
-	usageStateStore           *usageStateStore
+}
+
+// isNamespaceExcluded checks if a namespace should be ignored by the controller.
+// It checks if the namespace is the controller's own namespace, in the excluded list, or has the exclusion label.
+func (r *ClusterResourceQuotaReconciler) isNamespaceExcluded(ns *corev1.Namespace) bool {
+	if slices.Contains(r.ExcludedNamespaces, ns.Name) {
+		return true
+	}
+	if r.ExcludeNamespaceLabelKey == "" {
+		return false
+	}
+	_, hasLabel := ns.Labels[r.ExcludeNamespaceLabelKey]
+	return hasLabel
 }
 
 // Reconcile is part of the main kubernetes reconciliation loop which aims to
@@ -122,9 +232,6 @@ func (r *ClusterResourceQuotaReconciler) Reconcile(ctx context.Context, req ctrl
 
 	// Check for namespace changes and emit events
 	r.handleNamespaceChanges(crq, selectedNamespaces)
-	if r.usageStateStore != nil {
-		r.usageStateStore.ensureQuotaNamespaces(crq.Name, selectedNamespaces)
-	}
 
 	r.logger.Debug("Found namespaces matching selection criteria",
 		zap.Int("count", len(selectedNamespaces)),
@@ -190,6 +297,321 @@ func (r *ClusterResourceQuotaReconciler) Reconcile(ctx context.Context, req ctrl
 
 	metrics.QuotaReconcileTotal.WithLabelValues(crq.Name, "success").Inc()
 	return ctrl.Result{}, nil
+}
+
+// calculateAndAggregateUsage calculates the current resource usage for the given CRQ.
+func (r *ClusterResourceQuotaReconciler) calculateAndAggregateUsage(
+	ctx context.Context,
+	crq *quotav1alpha1.ClusterResourceQuota,
+	namespaces []string,
+) (quotav1alpha1.ResourceList, []quotav1alpha1.ResourceQuotaStatusByNamespace, error) {
+	r.logger.Debug("Calculating resource usage", zap.String("crq", crq.Name))
+	timer := prometheus.NewTimer(metrics.QuotaAggregationDuration.WithLabelValues(crq.Name))
+	defer timer.ObserveDuration()
+
+	totalUsage := make(quotav1alpha1.ResourceList)
+	usageByNamespace := make([]quotav1alpha1.ResourceQuotaStatusByNamespace, len(namespaces))
+	nsIndexMap := make(map[string]int)
+
+	resourceSnapshots := map[string]namespaceResourceSnapshot{}
+	if shouldPrefetchNamespaceResources(crq.Spec.Hard) {
+		var err error
+		resourceSnapshots, err = r.prefetchNamespaceResources(ctx, namespaces)
+		if err != nil {
+			r.logger.Warn(
+				"Failed to prefetch namespace resources, falling back to calculators",
+				zap.Error(err),
+				zap.String("crq", crq.Name),
+			)
+			resourceSnapshots = nil
+		}
+	}
+
+	// Initialize maps for efficient lookup
+	for i, nsName := range namespaces {
+		usageByNamespace[i] = quotav1alpha1.ResourceQuotaStatusByNamespace{
+			Namespace: nsName,
+			Status: quotav1alpha1.ResourceQuotaStatus{
+				Used: make(quotav1alpha1.ResourceList),
+			},
+		}
+		nsIndexMap[nsName] = i
+	}
+
+	// Iterate over each resource defined in the CRQ spec
+	for resourceName := range crq.Spec.Hard {
+		// Initialize total usage for this resource
+		totalUsage[resourceName] = resource.Quantity{}
+
+		// Detect storage class–scoped quota resources
+		resourceStr := string(resourceName)
+		if strings.HasSuffix(resourceStr, ".storageclass.storage.k8s.io/requests.storage") {
+			// Example: fast-ssd.storageclass.storage.k8s.io/requests.storage
+			storageClass := strings.TrimSuffix(resourceStr, ".storageclass.storage.k8s.io/requests.storage")
+			for _, nsName := range namespaces {
+				var currentUsage resource.Quantity
+				stepStart := time.Now()
+				if snapshot, ok := resourceSnapshots[nsName]; ok {
+					currentUsage = calculateStorageClassUsageFromPVCs(snapshot.PVCs, storageClass)
+				} else if r.StorageCalculator != nil {
+					storageUsage, err := r.StorageCalculator.CalculateStorageClassUsage(ctx, nsName, storageClass)
+					if err != nil {
+						metrics.QuotaAggregationStepDuration.WithLabelValues(crq.Name, "storage_class_usage").Observe(time.Since(stepStart).Seconds())
+						return nil, nil, fmt.Errorf("failed to calculate storage class usage for %s in %s: %w",
+							storageClass, nsName, err)
+					}
+					currentUsage = storageUsage
+				} else {
+					r.logger.Error("StorageCalculator is nil",
+						zap.String("namespace", nsName), zap.Stringer("resource", resourceName))
+					currentUsage = resource.MustParse("0")
+				}
+				metrics.QuotaAggregationStepDuration.WithLabelValues(crq.Name, "storage_class_usage").Observe(time.Since(stepStart).Seconds())
+				nsIndex := nsIndexMap[nsName]
+				usageByNamespace[nsIndex].Status.Used[resourceName] = currentUsage
+				if existing, exists := totalUsage[resourceName]; exists {
+					existing.Add(currentUsage)
+					totalUsage[resourceName] = existing
+				} else {
+					totalUsage[resourceName] = currentUsage
+				}
+			}
+			continue
+		}
+		if strings.HasSuffix(resourceStr, ".storageclass.storage.k8s.io/persistentvolumeclaims") {
+			// Example: fast-ssd.storageclass.storage.k8s.io/persistentvolumeclaims
+			storageClass := strings.TrimSuffix(resourceStr, ".storageclass.storage.k8s.io/persistentvolumeclaims")
+			for _, nsName := range namespaces {
+				var currentCount int64
+				stepStart := time.Now()
+				if snapshot, ok := resourceSnapshots[nsName]; ok {
+					currentCount = calculateStorageClassCountFromPVCs(snapshot.PVCs, storageClass)
+				} else if r.StorageCalculator != nil {
+					count, err := r.StorageCalculator.CalculateStorageClassCount(ctx, nsName, storageClass)
+					if err != nil {
+						metrics.QuotaAggregationStepDuration.WithLabelValues(crq.Name, "storage_class_count").Observe(time.Since(stepStart).Seconds())
+						return nil, nil, fmt.Errorf("failed to calculate storage class PVC count for %s in %s: %w",
+							storageClass, nsName, err)
+					}
+					currentCount = count
+				} else {
+					r.logger.Error("StorageCalculator is nil",
+						zap.String("namespace", nsName), zap.Stringer("resource", resourceName))
+					currentCount = 0
+				}
+				metrics.QuotaAggregationStepDuration.WithLabelValues(crq.Name, "storage_class_count").Observe(time.Since(stepStart).Seconds())
+				nsIndex := nsIndexMap[nsName]
+				usageByNamespace[nsIndex].Status.Used[resourceName] = *resource.NewQuantity(currentCount, resource.DecimalSI)
+				if existing, exists := totalUsage[resourceName]; exists {
+					existing.Add(*resource.NewQuantity(currentCount, resource.DecimalSI))
+					totalUsage[resourceName] = existing
+				} else {
+					totalUsage[resourceName] = *resource.NewQuantity(currentCount, resource.DecimalSI)
+				}
+			}
+			continue
+		}
+
+		for _, nsName := range namespaces {
+			// If nsName is empty, skip usage calculation for this entry
+			if nsName == "" {
+				r.logger.Info("Skipping usage calculation for empty namespace name")
+				continue
+			}
+			stepName := r.aggregationStepForResource(resourceName)
+			stepStart := time.Now()
+
+			currentUsage, calcErr := r.resolveNamespaceResourceUsage(ctx, nsName, resourceName, resourceSnapshots)
+
+			if calcErr != nil {
+				metrics.QuotaAggregationStepDuration.WithLabelValues(crq.Name, stepName).Observe(time.Since(stepStart).Seconds())
+				return nil, nil, calcErr
+			}
+
+			metrics.QuotaAggregationStepDuration.WithLabelValues(crq.Name, stepName).Observe(time.Since(stepStart).Seconds())
+
+			// Update usage for the specific namespace
+			nsIndex := nsIndexMap[nsName]
+			usageByNamespace[nsIndex].Status.Used[resourceName] = currentUsage
+
+			// Aggregate total usage correctly
+			// Since resource.Quantity has pointer receiver methods, we need to be careful
+			// about how we handle the aggregation
+			if existing, exists := totalUsage[resourceName]; exists {
+				existing.Add(currentUsage)
+				totalUsage[resourceName] = existing
+			} else {
+				totalUsage[resourceName] = currentUsage
+			}
+		}
+	}
+
+	r.logger.Debug("Usage calculation finished.")
+	return totalUsage, usageByNamespace, nil
+}
+
+func (r *ClusterResourceQuotaReconciler) resolveNamespaceResourceUsage(
+	ctx context.Context,
+	nsName string,
+	resourceName corev1.ResourceName,
+	resourceSnapshots map[string]namespaceResourceSnapshot,
+) (resource.Quantity, error) {
+	snapshot, hasSnapshot := resourceSnapshots[nsName]
+
+	switch resourceName {
+	case corev1.ResourceRequestsCPU,
+		corev1.ResourceRequestsMemory,
+		corev1.ResourceLimitsCPU,
+		corev1.ResourceLimitsMemory,
+		corev1.ResourcePods:
+		if hasSnapshot {
+			return calculateComputeUsageFromPods(snapshot.Pods, resourceName), nil
+		}
+		return r.calculateComputeResources(ctx, nsName, resourceName)
+	case corev1.ResourceRequestsStorage:
+		if hasSnapshot {
+			return calculateStorageUsageFromPVCs(snapshot.PVCs, resourceName), nil
+		}
+		return r.calculateStorageResources(ctx, nsName, resourceName)
+	case usage.ResourcePersistentVolumeClaims:
+		if hasSnapshot {
+			return calculatePVCCountUsageFromPVCs(snapshot.PVCs), nil
+		}
+		return r.calculateStorageResources(ctx, nsName, resourceName)
+	case usage.ResourceServices, usage.ResourceServicesLoadBalancers, usage.ResourceServicesNodePorts:
+		if hasSnapshot {
+			return calculateServiceUsageFromServices(snapshot.Services, resourceName), nil
+		}
+		return r.calculateServiceResources(ctx, nsName, resourceName)
+	default:
+		// Handle extended resources (hugepages, GPUs, etc.) via compute calculator.
+		if r.isComputeResource(resourceName) {
+			if hasSnapshot {
+				return calculateComputeUsageFromPods(snapshot.Pods, resourceName), nil
+			}
+			return r.calculateComputeResources(ctx, nsName, resourceName)
+		}
+		return r.calculateObjectCount(ctx, nsName, resourceName)
+	}
+}
+
+func shouldPrefetchNamespaceResources(hard quotav1alpha1.ResourceList) bool {
+	for resourceName := range hard {
+		resourceStr := string(resourceName)
+		switch resourceName {
+		case corev1.ResourceRequestsCPU,
+			corev1.ResourceRequestsMemory,
+			corev1.ResourceLimitsCPU,
+			corev1.ResourceLimitsMemory,
+			corev1.ResourcePods,
+			corev1.ResourceRequestsStorage,
+			usage.ResourcePersistentVolumeClaims,
+			usage.ResourceServices,
+			usage.ResourceServicesLoadBalancers,
+			usage.ResourceServicesNodePorts:
+			return true
+		}
+
+		if strings.HasSuffix(resourceStr, ".storageclass.storage.k8s.io/requests.storage") ||
+			strings.HasSuffix(resourceStr, ".storageclass.storage.k8s.io/persistentvolumeclaims") {
+			return true
+		}
+	}
+
+	return false
+}
+
+func calculateComputeUsageFromPods(pods []corev1.Pod, resourceName corev1.ResourceName) resource.Quantity {
+	return pod.CalculateUsageFromPods(pods, resourceName)
+}
+
+func calculateServiceUsageFromServices(svcs []corev1.Service, resourceName corev1.ResourceName) resource.Quantity {
+	return services.CalculateUsageFromServices(svcs, resourceName)
+}
+
+func calculateStorageUsageFromPVCs(pvcs []corev1.PersistentVolumeClaim, resourceName corev1.ResourceName) resource.Quantity {
+	return storage.CalculateStorageUsageFromPVCs(pvcs, resourceName)
+}
+
+func calculatePVCCountUsageFromPVCs(pvcs []corev1.PersistentVolumeClaim) resource.Quantity {
+	return storage.CalculatePVCCountUsageFromPVCs(pvcs)
+}
+
+func calculateStorageClassUsageFromPVCs(pvcs []corev1.PersistentVolumeClaim, storageClass string) resource.Quantity {
+	return storage.CalculateStorageClassUsageFromPVCs(pvcs, storageClass)
+}
+
+func calculateStorageClassCountFromPVCs(pvcs []corev1.PersistentVolumeClaim, storageClass string) int64 {
+	return storage.CalculateStorageClassCountFromPVCs(pvcs, storageClass)
+}
+
+func pvcMatchesStorageClass(pvc *corev1.PersistentVolumeClaim, storageClass string) bool {
+	return storage.PVCMatchesStorageClass(pvc, storageClass)
+}
+
+type namespaceResourceSnapshot struct {
+	Pods     []corev1.Pod
+	Services []corev1.Service
+	PVCs     []corev1.PersistentVolumeClaim
+}
+
+func (r *ClusterResourceQuotaReconciler) prefetchNamespaceResources(
+	ctx context.Context,
+	namespaces []string,
+) (map[string]namespaceResourceSnapshot, error) {
+	if r.KubeClient == nil {
+		return nil, fmt.Errorf("kube client is nil")
+	}
+
+	snapshots := make(map[string]namespaceResourceSnapshot, len(namespaces))
+	for _, nsName := range namespaces {
+		if nsName == "" {
+			continue
+		}
+
+		pods, err := r.KubeClient.CoreV1().Pods(nsName).List(ctx, metav1.ListOptions{})
+		if err != nil {
+			return nil, fmt.Errorf("failed to prefetch pods in namespace %s: %w", nsName, err)
+		}
+
+		svcs, err := r.KubeClient.CoreV1().Services(nsName).List(ctx, metav1.ListOptions{})
+		if err != nil {
+			return nil, fmt.Errorf("failed to prefetch services in namespace %s: %w", nsName, err)
+		}
+
+		pvcs, err := r.KubeClient.CoreV1().PersistentVolumeClaims(nsName).List(ctx, metav1.ListOptions{})
+		if err != nil {
+			return nil, fmt.Errorf("failed to prefetch pvcs in namespace %s: %w", nsName, err)
+		}
+
+		snapshots[nsName] = namespaceResourceSnapshot{
+			Pods:     pods.Items,
+			Services: svcs.Items,
+			PVCs:     pvcs.Items,
+		}
+	}
+
+	return snapshots, nil
+}
+
+func (r *ClusterResourceQuotaReconciler) aggregationStepForResource(resourceName corev1.ResourceName) string {
+	switch resourceName {
+	case corev1.ResourceRequestsCPU,
+		corev1.ResourceRequestsMemory,
+		corev1.ResourceLimitsCPU,
+		corev1.ResourceLimitsMemory,
+		corev1.ResourcePods:
+		return "compute"
+	case corev1.ResourceRequestsStorage:
+		return "storage"
+	case usage.ResourceServices, usage.ResourceServicesLoadBalancers, usage.ResourceServicesNodePorts:
+		return "services"
+	default:
+		if r.isComputeResource(resourceName) {
+			return "compute_extended"
+		}
+		return "object_count"
+	}
 }
 
 // calculateObjectCount calculates the usage for object count quotas.
@@ -288,6 +710,67 @@ func (r *ClusterResourceQuotaReconciler) updateStatus(
 	return r.Status().Patch(ctx, crqCopy, client.MergeFrom(crq))
 }
 
+// findQuotasForObject maps objects (including Namespaces and other namespaced resources) to ClusterResourceQuota requests
+// that should be reconciled based on namespace selection criteria. This unified function handles both:
+// - Namespace objects directly (when namespaces are created, updated, or deleted)
+// - Other namespaced objects (Pods, Services, etc.) by first retrieving their namespace
+// It excludes the controller's own namespace and any namespaces marked with the exclusion label.
+func (r *ClusterResourceQuotaReconciler) findQuotasForObject(ctx context.Context, obj client.Object) []reconcile.Request {
+	// Handle nil object gracefully
+	if obj == nil {
+		return nil
+	}
+
+	var ns *corev1.Namespace
+	var err error
+
+	// Handle Namespace objects directly
+	if namespace, ok := obj.(*corev1.Namespace); ok {
+		ns = namespace
+	} else {
+		// For cluster-scoped resources, return nil (no quota mapping needed)
+		namespaceName := obj.GetNamespace()
+		if namespaceName == "" {
+			return nil
+		}
+
+		// For other objects, get the namespace they belong to
+		ns = &corev1.Namespace{}
+		if err = r.Get(ctx, types.NamespacedName{Name: namespaceName}, ns); err != nil {
+			r.logger.Error("Failed to get namespace for object to check for exclusion", zap.Error(err), zap.String("object", client.ObjectKeyFromObject(obj).String()))
+			return nil
+		}
+	}
+
+	if r.isNamespaceExcluded(ns) {
+		return nil // Ignore events from excluded namespaces
+	}
+
+	// Find which CRQ selects this namespace.
+	crq, err := r.crqClient.GetCRQByNamespace(ctx, ns)
+	if err != nil {
+		r.logger.Error("Failed to get ClusterResourceQuota for namespace", zap.Error(err))
+		return nil
+	}
+	if crq != nil {
+		r.logger.Debug("Found ClusterResourceQuota for namespace", zap.String("crq", crq.Name), zap.String("namespace", ns.Name))
+	} else {
+		r.logger.Debug("No ClusterResourceQuota found for namespace", zap.String("namespace", ns.Name))
+	}
+
+	if crq != nil {
+		return []reconcile.Request{
+			{
+				NamespacedName: types.NamespacedName{
+					Name: crq.Name,
+				},
+			},
+		}
+	}
+
+	return nil
+}
+
 // isComputeResource determines if a resource type should be calculated using the compute calculator.
 // This includes standard compute resources and extended resources (hugepages, GPUs, etc.)
 func (r *ClusterResourceQuotaReconciler) isComputeResource(resourceName corev1.ResourceName) bool {
@@ -312,4 +795,133 @@ func (r *ClusterResourceQuotaReconciler) isComputeResource(resourceName corev1.R
 
 	// If we can't categorize it, assume it's not a compute resource
 	return false
+}
+
+// SetupWithManager sets up the controller with the Manager.
+func (r *ClusterResourceQuotaReconciler) SetupWithManager(ctx context.Context, cfg *config.Config, mgr ctrl.Manager) error {
+	// Initialize logger
+	if r.logger == nil {
+		r.logger = zap.L().Named("clusterresourcequota-controller")
+	}
+
+	// Initialize the KubeClient using the manager's config if not already set
+	if r.KubeClient == nil {
+		cfg := mgr.GetConfig()
+		r.KubeClient = kubernetes.NewForConfigOrDie(cfg)
+	}
+	k8sConfig := mgr.GetConfig()
+	clientset, err := kubernetes.NewForConfig(k8sConfig)
+	if err != nil {
+		return fmt.Errorf("unable to create kubernetes clientset: %w", err)
+	}
+	if r.crqClient == nil {
+		r.crqClient = quota.NewCRQClient(r.Client, r.logger)
+	}
+
+	if r.StorageCalculator == nil {
+		r.StorageCalculator = storage.NewStorageResourceCalculator(clientset, r.logger)
+	}
+	if r.ComputeCalculator == nil {
+		r.ComputeCalculator = pod.NewPodResourceCalculator(clientset, r.logger)
+	}
+	if r.ServiceCalculator == nil {
+		r.ServiceCalculator = services.NewServiceResourceCalculator(clientset, r.logger)
+	}
+	if r.ObjectCountCalculator == nil {
+		r.ObjectCountCalculator = objectcount.NewObjectCountCalculator(clientset, r.logger)
+	}
+
+	// Initialize EventRecorder
+	if r.EventRecorder == nil {
+		r.EventRecorder = events.NewEventRecorder(
+			mgr.GetEventRecorderFor("pac-quota-controller"),
+			cfg.OwnNamespace,
+			r.logger,
+		)
+	}
+
+	// Initialize previous namespaces tracking
+	if r.previousNamespacesByQuota == nil {
+		r.previousNamespacesByQuota = make(map[string][]string)
+	}
+
+	// Load event cleanup configuration from multiple sources
+	var cleanupConfig events.CleanupConfig
+
+	if r.Config != nil && r.Config.EventsEnable {
+		cleanupConfig, err = events.LoadEventCleanupConfig(
+			r.Config.EventsConfigPath,
+			r.Config.EventsTTL,
+			r.Config.EventsMaxEventsPerCRQ,
+			r.Config.EventsCleanupInterval,
+		)
+		if err != nil {
+			r.logger.Warn("Failed to load event cleanup config, using defaults", zap.Error(err))
+			cleanupConfig = events.DefaultCleanupConfig()
+		}
+	} else {
+		// Events disabled or no config provided, use defaults but disable cleanup
+		cleanupConfig = events.DefaultCleanupConfig()
+		if r.Config != nil && !r.Config.EventsEnable {
+			cleanupConfig.Enabled = false
+		}
+	}
+
+	cleanupManager := events.NewEventCleanupManager(mgr.GetClient(), cleanupConfig, r.logger)
+
+	// Start cleanup in background
+	go func() {
+		cleanupManager.Start(ctx)
+	}()
+
+	// Start periodic violation cache cleanup
+	go func() {
+		ticker := time.NewTicker(15 * time.Minute)
+		defer ticker.Stop()
+
+		for range ticker.C {
+			r.EventRecorder.CleanupExpiredViolations()
+		}
+	}()
+
+	r.logger.Info("Setting up ClusterResourceQuota controller")
+
+	// Predicate to filter out updates to status subresource
+	// This prevents reconcile loops caused by status updates
+	resourcePredicate := resourceUpdatePredicate{}
+
+	b := ctrl.NewControllerManagedBy(mgr).
+		For(&quotav1alpha1.ClusterResourceQuota{})
+
+	// Watch for changes to tracked resources and trigger reconciliation for associated CRQs
+	watchedObjectTypes := []struct {
+		obj   client.Object
+		preds []predicate.Predicate
+	}{
+		{&corev1.Namespace{}, nil},
+		{&corev1.Pod{}, []predicate.Predicate{resourcePredicate}},
+		{&corev1.PersistentVolumeClaim{}, nil},
+		{&corev1.Service{}, nil},
+		// Generic object count resources
+		{&corev1.ConfigMap{}, nil},
+		{&corev1.Secret{}, nil},
+		{&corev1.ReplicationController{}, nil},
+		{&appsv1.Deployment{}, nil},
+		{&appsv1.StatefulSet{}, nil},
+		{&appsv1.DaemonSet{}, nil},
+		{&batchv1.Job{}, nil},
+		{&batchv1.CronJob{}, nil},
+		{&autoscalingv1.HorizontalPodAutoscaler{}, nil},
+		{&networkingv1.Ingress{}, nil},
+	}
+	for _, w := range watchedObjectTypes {
+		b = b.Watches(
+			w.obj,
+			handler.EnqueueRequestsFromMapFunc(r.findQuotasForObject),
+			builder.WithPredicates(w.preds...),
+		)
+	}
+
+	return b.Named("clusterresourcequota").
+		Complete(r)
 }
